core: post-merge cleanup — rlib-only crate-type and clippy fixes
- Cargo.toml: drop cdylib from alephcore crate-type (no exported FFI
  consumer; it broke doctest transitive-dep resolution E0460/E0462)
- Fix clippy lints: redundant clones/casts, inspect over map, repeat_n,
  contains over == chain, doc-comment misplacement
- gateway: await async Orchestrator::end_session at both call sites
- doctor: bump REGISTERED_CHECKS to 14 for newly added check

+++ src/loop_graph/mod.rs
@@ -119,7 +119,7 @@ pub fn spawn_event_persister(
                     }
                 }
                 Err(tokio::sync::broadcast::error::RecvError::Lagged(n)) => {
-                    lag_counter.fetch_add(n as u64, std::sync::atomic::Ordering::AcqRel);
+                    lag_counter.fetch_add(n, std::sync::atomic::Ordering::AcqRel);
                     tracing::warn!(dropped = n,
                         "loop_graph: audit persister lagged — events dropped (audit may lose, never block)");
                 }
